requirements.txt.txt
openai
pinecone
neo4j
python-dotenv
sentence-transformers